+++ sdk/core/azure_core_amqp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.2.0-beta.2 (Unreleased)
+
+### Breaking Changes
+
+- Removed AMQP-over-WebSockets support, including the `AmqpTransport::WebSocket` variant and the `fe2o3_amqp_ws` and `fe2o3_amqp_ws_rustls` features, to eliminate the prohibited `sha1` dependency. `AmqpTransport::Tcp` and the APIs that select the transport remain available. ([#5261](https://github.com/Azure/azure-sdk-for-rust/issues/5261))
+
 ## 1.2.0-beta.1 (2026-09-04)
 
 ### Features Added